requirements.txt
Jinja2==2.8
Markdown==2.6.6
MarkupSafe==0.23
pkg-resources==0.0.0
Pygments==2.1.3
python-frontmatter==0.2.1
python-slugify==1.2.0
PyYAML==3.11
six==1.10.0
Unidecode==0.4.19
